Extract last block number param
